==> 00_1 Generate Dataset Data.ipynb <==
# Import Stuff

import sys

from Modules import generate_kallisto_data
from Modules import generate_merged_data
from Modules import generate_alignment_statistics

# Create Kallisto Files

#generate_kallisto_data.get_kallisto_data('Lab_Pvalb', skip_types=[])
#generate_kallisto_data.get_kallisto_data('Lin_New', skip_types=[])
#generate_kallisto_data.get_kallisto_data('Lab_Cortex', skip_types=[])
#generate_kallisto_data.get_kallisto_data('Lab_OLM', skip_types=['non-OLM'])
#generate_kallisto_data.get_kallisto_data('GSE99888', skip_types=['Unknown'])
#generate_kallisto_data.get_kallisto_data('GSE75386')
#generate_kallisto_data.get_kallisto_data('GSE119248')
#generate_kallisto_data.get_kallisto_data('GSE70844')
#generate_kallisto_data.get_kallisto_data('Cadwell')
#generate_kallisto_data.get_kallisto_data('GSE60361')
#generate_kallisto_data.get_kallisto_data('GSE109999')

# Get TPM Data

#generate_merged_data.generate_tpm_data('Lab_Pvalb', QC=True, skip_types=['DEL', 'UC', 'Unknown', 'Cortical'])
#generate_merged_data.generate_tpm_data('Lab_Pvalb', QC=False, skip_types=['DEL', 'Unknown'], savename='Pvalb_All')
#generate_merged_data.generate_tpm_data('Lin_New', QC=True)
#generate_merged_data.generate_tpm_data('Lab_Cortex', QC=True, QC_ref='Lab_Pvalb')
#generate_merged_data.generate_tpm_data('Lab_Pvalb_Facs_02', QC=True)
#generate_merged_data.generate_tpm_data('Lab_OLM', QC=False)
#generate_merged_data.generate_tpm_data('GSE99888', QC=False)
#generate_merged_data.generate_tpm_data('GSE75386')
#generate_merged_data.generate_tpm_data('GSE119248')
#generate_merged_data.generate_tpm_data('GSE70844')
#generate_merged_data.generate_tpm_data('Cadwell')
#generate_merged_data.generate_tpm_data('Lab_Gabe', level=1)
#generate_merged_data.generate_tpm_data('GSE60361')
#generate_merged_data.generate_tpm_data('Gauwens')
#generate_merged_data.generate_tpm_data('GSE109999', QC=True)

# Get Alignment Quality Statistics

#generate_alignment_statistics.generate_alignment_stats('Lab_Pvalb')
#generate_alignment_statistics.generate_alignment_stats('Lin_New')
#generate_alignment_statistics.generate_alignment_stats('Lab_Cortex')
#generate_alignment_statistics.generate_alignment_stats('Lab_OLM')
#generate_alignment_statistics.generate_alignment_stats('GSE75386')
#generate_alignment_statistics.generate_alignment_stats('GSE99888')
#generate_alignment_statistics.generate_alignment_stats('GSE119248')
#generate_alignment_statistics.generate_alignment_stats('GSE70844')
#generate_alignment_statistics.generate_alignment_stats('Cadwell')
#generate_alignment_statistics.generate_alignment_stats('GSE60361')
#generate_alignment_statistics.generate_alignment_stats('Gauwens')
#generate_alignment_statistics.generate_alignment_stats('GSE109999')

# Get Loom Data

params = {'age_cutoff':20, 'min_count':3, 'log':True, 'ending':'tpm', 'markers':False}
#generate_merged_data.generate_loompy_data(['Lab_Pvalb', 'Lab_OLM'], 'Lab_Data', **params)
#generate_merged_data.generate_loompy_data(['Lab_Pvalb'], 'Lab_Pvalb', **params)
params['markers'] = True
#generate_merged_data.generate_loompy_data(['Lab_Pvalb'], 'Lab_Pvalb_144', **params)
#generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris', age_cutoff=None, log=False)
'''celltypes = ['Pvalb.C1ql1.Cpne5', 'Pvalb.C1ql1.Npy', 'Pvalb.C1ql1.Pvalb',
             'Pvalb.Tac1.Akr1c18', 'Pvalb.Tac1.Nr4a2', 'Pvalb.Tac1.Sst',
             'Pvalb.Tac1.Syt2', 'Sst.Erbb4.Crh', 'Sst.Erbb4.Rgs10',
             'Sst.Erbb4.Th']
#generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_trim', age_cutoff=None, log=False, celltypes=celltypes, markers=False)
#generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_trim_log', age_cutoff=None, log=True, celltypes=celltypes, markers=False)

celltypes = ['Pvalb.C1ql1.Cpne5', 'Pvalb.C1ql1.Npy', 'Pvalb.C1ql1.Pvalb',
             'Pvalb.Tac1.Akr1c18', 'Pvalb.Tac1.Nr4a2', 'Pvalb.Tac1.Sst',
             'Pvalb.Tac1.Syt2']
generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_Pvalb', age_cutoff=None, log=True, celltypes=celltypes, markers=False)

celltypes = ['Pvalb.C1ql1.Cpne5', 'Pvalb.C1ql1.Npy', 'Pvalb.C1ql1.Pvalb']
generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_C1ql1', age_cutoff=None, log=True, celltypes=celltypes, markers=False)

celltypes = ['Pvalb.Tac1.Akr1c18', 'Pvalb.Tac1.Nr4a2', 'Pvalb.Tac1.Sst',
             'Pvalb.Tac1.Syt2', 'Sst.Erbb4.Crh', 'Sst.Erbb4.Rgs10',
             'Sst.Erbb4.Th']
generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_Tac1Erbb4', age_cutoff=None, log=True, celltypes=celltypes, markers=False)

celltypes = ['Pvalb.Tac1.Akr1c18', 'Pvalb.Tac1.Nr4a2', 'Pvalb.Tac1.Sst',
             'Pvalb.Tac1.Syt2']
generate_merged_data.generate_loompy_data(['GSE99888'], 'Lab_Harris_Tac1', age_cutoff=None, log=True, celltypes=celltypes, markers=False)'''

pass
