==> 02_2 EdgeR Comparisons.ipynb <==
suppressMessages({
    library(SingleCellExperiment)
    #library(zinbwave)
    library(scRNAseq)
    library(matrixStats)
    library(magrittr)
    library(edgeR)
    library(scran)
})

get_dataset <- function(dataset){
    fname <- sprintf('Datasets/%s_edgeR.RData', dataset)
    dge <- readRDS(fname)
    
    return (dge)
}

get_cell_labels <- function(dataset){
    fname <- sprintf('Datasets/%s-transcriptional_labels.tsv', dataset)
    cell_labels <- read.table(fname, sep='\t', header=1, row=1)
    
    return (cell_labels)
}

get_weighted_vars <- function(dge, ctype1, ctype2){
    ctypes <- dge$samples$group
    num1 = sum(ctypes %in% ctype1)
    num2 = sum(ctypes %in% ctype2)
    weights = rep(num1, length(ctypes))
    weights[ctypes %in% ctype1] = num2

    wvars <- rowWeightedVars(cpm(dge) %>% log1p, w=weights)
    
    return (wvars)
}

calc_weighted_vars <- function(dge, ctype1, ctype2){
    ctypes <- dge$samples$group
    count_mat <- cpm(dge)
    weights = rep(0, length(ctypes))
    for (ctype in unique(ctypes)){
        weights[ctypes==ctype] = 1 / sum(ctypes==ctype)
    }
    weights <- round(weights / min(weights))
    
    wvars <- rowWeightedVars(count_mat %>% log1p, w=weights)
    
    return (wvars)
}

filter_features <- function(dge, ctype1, ctype2){
    dge <- dge[,dge$samples$group %in% c(ctype1, ctype2)]
    
    #count1 = dim(dge[,dge$samples$group %in% ctype1])
    #count2 = dim(dge[,dge$samples$group %in% ctype2])
    #count = min(count1, count2)
    
    #cutoff_count = (count * 0.5)
    
    keep1 <- rowSums(cpm(dge)>15)> 4
    keep2 <- aveLogCPM(dge) > 4
    keep <- keep1 & keep2
    dge <- dge[keep,,keep.lib.sizes=FALSE]
    #cpm(dge) %>% log1p %>% rowVars -> vars
    vars <- get_weighted_vars(dge, ctype1, ctype2)
    names(vars) <- rownames(dge)
    vars <- sort(vars, decreasing = TRUE)
    #dge <- dge[names(vars)[1:5000],]
    
    return (dge)
}

run_comparison <- function(dge, ctype1, ctype2, fname){
    # get data of interest
    dge <- filter_features(dge, ctype1, ctype2)
    data = dge$samples
    data$group <- factor(data$group, levels=c(ctype1, ctype2))
    design <- model.matrix(~group, data = data)
    
    # run analysis
    dge <- estimateDisp(dge, design, robust=TRUE)
    #fit <- glmFit(dge, design, robust=TRUE)
    fit <- glmQLFit(dge, design, robust=TRUE)
    lrt <- glmQLFTest(fit, coef=2)
    #lrt <- glmWeightedF(fit, coef = 2)
    #lrt <- glmTreat(fit, coef=2, lfc=1)
    
    save_edgeR_result(lrt, fname)
}

run_multi_comparisons <- function(dge, ctypes1, ctypes2, fname){
    # get data of interest
    dge <- filter_features(dge, ctypes1, ctypes2)
    data <- dge$samples
    data$group <- factor(data$group, levels=c(levels(data$group), c('Type_1', 'Type_2')))
    data[data$group %in% ctypes1, 'group'] <- 'Type_1'
    data[data$group %in% ctypes2, 'group'] <- 'Type_2'
    data$group <- factor(data$group, levels=c('Type_1', 'Type_2'))
    design <- model.matrix(~group, data = data)
    
    # run analysis
    dge <- estimateDisp(dge, design)
    fit <- glmQLFit(dge, design, robust=TRUE)
    lrt <- glmQLFTest(fit, coef=2)
    
    save_edgeR_result(lrt, fname)
}

save_edgeR_result <- function(lrt, fname){
    txtname <- sprintf('Differential/edgeR/%s.txt', fname)
    df <- topTags(lrt, n=dim(lrt$table)[1])$table
    df <- df[order(df$PValue), ]
    write.table(df, file=txtname, sep='\t', quote=FALSE)
}

run_all_comparisons <- function(dge, labels, folder){
    dge$samples$group <- labels
    group = as.character(sort(unique(labels)))
    
    for (i in 1:(length(group)-1)){
        for (j in (i+1):length(group)){
            ctype1 <- group[[i]]
            ctype2 <- group[[j]]
            title <- sprintf('%s/%s vs %s', folder, ctype1, ctype2)
            run_comparison(dge, ctype1, ctype2, title)
        }
    }
}

dge <- get_dataset('Lab_Pvalb')
cell_labels <- get_cell_labels('Lab_Pvalb')
cell_labels <- cell_labels[colnames(dge),]
cell_labels$Morph_Gender.PV.types <- paste(cell_labels$Morph.PV.types, cell_labels$Gender.PV.types, sep='_')
cell_labels$Morph_Side.PV.types <- paste(cell_labels$Morph.PV.types, cell_labels$Side.PV.types, sep='_')

labels <- cell_labels$Morph.PV.types
run_all_comparisons(dge, labels, 'Morph')
labels <- cell_labels$proMMT.PV.types
run_all_comparisons(dge, labels, 'proMMT')

run_multi_comparisons(dge, 'vAAC', c('vBC', 'hBC'), 'MorphMarker/AAC vs BC')
run_multi_comparisons(dge, 'vAAC', c('vBIC', 'hBIC'), 'MorphMarker/AAC vs BIC')
run_multi_comparisons(dge, c('vBC', 'hBC'), c('vBIC', 'hBIC'), 'MorphMarker/BC vs BIC')
run_multi_comparisons(dge, c('vAAC', 'vBC', 'vBIC'), c('hBC', 'hBIC'), 'MorphDirectional/Vertical vs Horizontal')

dge$samples$group <- cell_labels$Mapping.PV.types
run_comparison(dge, 'Continent 2', 'Continent 3', 'Mapping/Pvalb.Tac1 vs Pvalb.C1ql1')

dge <- get_dataset('Harris_Pvalb')
cell_labels <- get_cell_labels('Harris_Pvalb')
cell_labels$Group <- 'Pvalb.Tac1'
cell_labels[cell_labels$CellType %in% c('Pvalb.C1ql1.Pvalb', 'Pvalb.C1ql1.Cpne5', 'Pvalb.C1ql1.Npy'), 'Group'] <- 'Pvalb.C1ql1'

run_all_comparisons(dge, cell_labels$Group, 'Harris_Pvalb/Continent')
